some code clean ups
